virtio-blk: add virtual pid to block layer
I/O scheduler in the host use the PID to group request before sending
them to the disk device driver. This patch enables passing the PID
from the guest process (Virtual PID) to the host.

If you use disk images for your guest which are stored on a hard-drive,
this will get better throughput and a better behaviour of CFQ on your
guest workload. Beware that you need to have a virtual machine monitor
which supports this kind of information.

Signed-off-by: Matthias Brugger <matthias.bgg@gmail.com>