FROMLIST: platform/chrome: cros_ec_trace: Match trace commands with EC commands
There are some EC commands that are not included yet as trace commands,
in order to get all the traces for the all supported commands match the
commands accordingly.

Note that a change, adding or removing an EC command, should be
reflected in the cros_ec_trace.c file in order to avoid mismatches
again.

The list of current commands is generated using the following script:

 sed -n 's/^#define \(EC_CMD_[[:alnum:]_]*\)\s.*/\tTRACE_SYMBOL(\1),\\/p' \
        include/linux/platform_data/cros_ec_commands.h

Signed-off-by: Enric Balletbo i Serra <enric.balletbo@collabora.com>
Series-version: 2
Series-to: lkml
Series-cc: chrome-platform, collabora
Series-cc: tzungbi@google.com, rrangel@chromium.org, gwendal@chromium.org
Signed-off-by: Enric Balletbo i Serra <enric.balletbo@collabora.com>