FROMLIST: platform/chrome: cros_ec: Match implementation with headers
The 'cros_ec' core driver is the common interface for the cros_ec
transport drivers to do the shared operations to register, unregister,
suspend and resume. The interface is provided by including the header
'include/linux/platform_data/cros_ec_proto.h', however, instead of have
the implementation of these functions in cros_ec_proto.c, it is in
'cros_ec.c', which is a different kernel module. Apart from being a bad
practice, this can induce confusions allowing the users of the cros_ec
protocol to call these functions.

The register, unregister, suspend and resume functions *should* only be
called by the different transport drivers (i2c, spi, lpc, etc.), so make
this a bit less confusing by moving these functions from the public
in-kernel space to a private include in platform/chrome, and then, the
interface for cros_ec module and for the cros_ec_proto module is clean.

Signed-off-by: Enric Balletbo i Serra <enric.balletbo@collabora.com>
Series-to: lkml
Series-cc: chrome-platform, collabora