FROMLIST: drm/bridge: ps8640: Return an error for incorrect attach flags
Bridge drivers that implement the new model only shall return an error
from their attach() handler when the DRM_BRIDGE_ATTACH_NO_CONNECTOR flag
is not set. So make sure we return an error because only the new
drm_bridge model is supported.

Signed-off-by: Enric Balletbo i Serra <enric.balletbo@collabora.com>
Series-to: lkml
Series-cc: collabora, mediatek, laurent.pinchart@ideasonboard.com, matthias.bgg@gmail.com, sam@ravnborg.org
Cover-letter:
drm/bridge: ps8640: Make sure all needed is powered to get the EDID

The first patch is not really related and can be picked alone. The
second and the third are to rework the power handling to get the EDID.
Basically, we need to make sure all the needed is powered to be able
to get the EDID. Before, we saw that getting the EDID failed as
explained in the third patch.

Note that this series depends on:

  https://lore.kernel.org/patchwork/project/lkml/list/?series=448525

to apply cleanly.
END